[UPDATE] Refactor AppModule.kt
