+++ app/src/main/java/com/softyorch/taskapp/di/AppModule.kt
@@ -2,8 +2,11 @@ package com.softyorch.taskapp.di
 
 import android.content.Context
 import androidx.room.Room
+import com.softyorch.taskapp.data.database.datastore.DatastoreDataBase
 import com.softyorch.taskapp.data.database.tasks.TaskDatabase
 import com.softyorch.taskapp.data.database.tasks.TaskDatabaseDao
+import com.softyorch.taskapp.data.database.userdata.UserDataBase
+import com.softyorch.taskapp.data.database.userdata.UserDataBaseDao
 import com.softyorch.taskapp.utils.*
 import dagger.Module
 import dagger.Provides
@@ -29,4 +32,23 @@ object AppModule {
             TASK_TBL)
             .fallbackToDestructiveMigration()
             .build()
+
+    @Singleton
+    @Provides
+    fun providesUserData(userDataBase: UserDataBase): UserDataBaseDao = userDataBase.userDataDao()
+
+    @Singleton
+    @Provides
+    fun providesUserDataBase(@ApplicationContext context: Context): UserDataBase =
+        Room.databaseBuilder(
+            context,
+            UserDataBase::class.java,
+            USERDATA_TBL)
+            .fallbackToDestructiveMigration()
+            .build()
+
+    @Singleton
+    @Provides
+    fun providesDatastoreDataBase(@ApplicationContext context: Context): DatastoreDataBase =
+        DatastoreDataBase(context = context)
 }
