--- app/src/main/java/com/softyorch/taskapp/di/DataStoreModule.kt
@@ -1,20 +1,0 @@
-package com.softyorch.taskapp.di
-
-import android.content.Context
-import com.softyorch.taskapp.data.database.datastore.DatastoreDataBase
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.qualifiers.ApplicationContext
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@InstallIn(SingletonComponent::class)
-@Module
-object DataStoreModule {
-
-    @Singleton
-    @Provides
-    fun providesDatastoreDataBase(@ApplicationContext context: Context): DatastoreDataBase =
-        DatastoreDataBase(context = context)
-}
